requirements.txt
click==8.0.1
Flask==2.0.1
PyYaml==5.4.1